Add files via upload
This is a tool for building phrases. You have the option of building the phrase using abstracts (NOUN, VERB, etc.) and then filling it out, or just adding words raw. I have a few other ideas of how to make the tool more powerful or useful, but as it stands right now it should hopefully interface with our learning algorithm